Added date format to DBH addBike method

+++ src/myapp/data/Bike.java
@@ -14,7 +14,7 @@ public class Bike {
     private int totalTrips;
     private Location location;
 
-    public Bike(int price, Date purchased, String type, String make){
+    public Bike(double price, Date purchased, String type, String make) {
         this.price = price;
         this.purchased = purchased;
         this.type = type;
